Move banner entry
