+++ src/openrct2-ui/windows/Scenery.cpp
@@ -22,6 +22,7 @@
 #include <openrct2/network/network.h>
 #include <openrct2/object/ObjectList.h>
 #include <openrct2/object/WallObject.h>
+#include <openrct2/object/BannerObject.h>
 #include <openrct2/sprites.h>
 #include <openrct2/world/LargeScenery.h>
 #include <openrct2/world/Park.h>
