+++ src/openrct2-ui/windows/TopToolbar.cpp
@@ -53,6 +53,7 @@
 #include <openrct2/interface/Screenshot.h>
 #include <openrct2/localisation/Formatter.h>
 #include <openrct2/network/network.h>
+#include <openrct2/object/BannerObject.h>
 #include <openrct2/object/WallObject.h>
 #include <openrct2/paint/VirtualFloor.h>
 #include <openrct2/scenario/Scenario.h>
