+++ src/openrct2/paint/tile_element/Paint.Banner.cpp
@@ -14,6 +14,7 @@
 #include "../../interface/Viewport.h"
 #include "../../localisation/Formatter.h"
 #include "../../localisation/Localisation.h"
+#include "../../object/BannerObject.h"
 #include "../../profiling/Profiling.h"
 #include "../../ride/TrackDesign.h"
 #include "../../sprites.h"
